Frontend.
Added support for the server to scan the PDB before continuing with seqtol submission. This could be generalized to all protocols (requires the todo change in JS startup function).
Added the ranked table to the job info and data directory pages.
Allowed Cystein to be used in premutated residues for seqtol.
Fix seqtol kT value to 0.228.
Dropped the min recommended number of structures for seqtol.
Fixed UI bug where job could not be submitted if the user initially chose n chains then changed the choice to m chains where m < n.
Fixed UI bug where the Job Name and Number Of Structures fields were reset on choosing or changing the number of chains.

Todos:
Increase the rec. num of structures again.



git-svn-id: file:///kortemmelab/shared/commonSVN/rosettaweb@224 fc5780ad-6544-46e6-b60b-86c75d3988a5