Frontend:
PDBs are now saved by original filename in a subdirectory named by the session. The original name is stored in the database to avoid the long confusing names based on the SID.
Added a button to click when uploading a PDB by ID code. This is not necessary (the use can press enter) but may be more natural for some users.
Fixed the weight ordering for Sequence Tolerance so that it corresponds with the chain order in the PDB file in case the user enters the chains out of order.



git-svn-id: file:///kortemmelab/shared/commonSVN/rosettaweb@225 fc5780ad-6544-46e6-b60b-86c75d3988a5