Split protocol definition into base class and inherited classes for webserver and backend, allowing data to be shared between them and reused/modified easily. Both premutated and designed residues are now shown for the sequence tolerance protocol in the Jmol applet. Added database field to RosettaBinaries to centralize these settings. Fixed up daemon paths to new directory structure.
Deleted two unused scripts.
Added a compilation bug in rosettaseqtol_classic2.py (todo).
The PDB checker now suggests using mini only if that is an available option.

Fixed a bug in sequence tolerance storage introduced in recent refactoring (all_resids not defined).
Fixed a bug in getSortedString introduced when I started storing DB values are non-string values (int/float etc.).
Fixed a bug in the queue page based on the old "either mini or classic" logic.
Fixed a bug in the jobinfo page where we were performing a time calculation on a None value (introduced in a recent commit).
Fixed a bug where you could never submit a job if the classic PDB check failed. Related change is that the binary selector is shown in the first submission stage and a textbox has been added.
Fixed a bug in JS PDB filename recognition and in a use of checkValue (its return type is undefined/int, not boolean).



git-svn-id: file:///kortemmelab/shared/commonSVN/rosettaweb@246 fc5780ad-6544-46e6-b60b-86c75d3988a5

Former-commit-id: c1f965d40e01d14e1a9eb0578794f95f5e552c41