Backend: Fixed a bug (in Python, I would claim) writing resfiles with residue ids of 0. Fixed a bug in check_files.
Frontend:
Fixed a bug where the wrong binary was being passed to the backend (ChosenBinary).
Fixed a bug with Ensemble db storage (a table was expected).
Added Tanja's new text.
Change the colour of the arrow pointing to the protocol to make it more obvious which protocol is being used (protocolarrowi_j).
Added a recommendation for mini for point mutations and changed default to mini.
Fixed the citation for the new seqtol paper.
Seqtol logo change (to be redone).



git-svn-id: file:///kortemmelab/shared/commonSVN/rosettaweb@271 fc5780ad-6544-46e6-b60b-86c75d3988a5

Former-commit-id: 297fbada82a0a03c3f140e10cf185589d3474eb4