Nice update for cluster code. Note: Debug mode has been turned on in this commit.
An SGEConnection class has been added and shared between jobs, allowing us to cache qstat results and only call qstat from the controller (daemon or cluster.py) when allowed. This prevents us from bombarding the head node with requests and allows us to sleep an appropriate amount of time in the main controller loop.

This class provides useful functions such as a qstat summary organised by database PKs and a diff function. This diff function is now used to only print out status changes rather than a continuous stream of repeated statuses which will keep the log file smaller.

We now delete the temporary directories on netapp and the webserver in at least some cases.

Tasks can now call qdel and delete themselves from the cluster on failure rather than us having to do it manually.

We now keep track of the database PK for any jobs recently submitted (last 200) so that the same job cannot get submitted to the server repeatedly in the case where the controlling logic fails. This shouldn't happen but the safeguard is reassuring.

A bug in recordErrorInJob has been fixed so job failures are now handled properly and update the DB.

Removed some output regarding copying files used for debugging. Added ids (DB PK, task ID) to some output messages to identify whence they came.

Laid some foundations for plan to generate graphs of a job status; jobs now remember their initial tasks and there is a crude traversal function.

todo: Log CachedTable to file on qstat update so we can kill the jobs on daemon restart if they are still alive (also do this at the entry point if there is a terminal failure).

git-svn-id: file:///kortemmelab/shared/commonSVN/rosettaweb@286 fc5780ad-6544-46e6-b60b-86c75d3988a5

Former-commit-id: 3a1dd84b367837852c9c3ddc307239b9781b659a