Added a text file with instructions for testing the webpage.
Changes to text from meeting with Tanja.
Fixed a couple of bugs in EnsembleDesign datadir.
Changed PDBID maxlength to fix oddly reported Chrome bug when JS tried to set its value to a string longer than maxlength.
Turned off cluster progress graph when it doesn't exist yet.
Fixed a historic bug in _show_scores_file which relied on string comparison rather than float comparison. _getPDBfiles does not sort the list of files anymore either so Jmol should get them in the correct order. I'll be testing this now.
Fixed bug in seqtolHK show function where the PDBs were not listed correctly and the designed residues were not shown. Fixed the other links in here as well. The residues are now displayed as designed rather than mutated.

todo: I started to fix a bug involving resubmitting the same job on Albana but it needs to be finished.



git-svn-id: file:///kortemmelab/shared/commonSVN/rosettaweb@332 fc5780ad-6544-46e6-b60b-86c75d3988a5

Former-commit-id: 72dd33f6b1171971e4bfd29fbf6de3aa0676d44c