Daemon: Removed some commented out code dealing with the old locking logic.
Frontend:
Fixed up HTML links for remote jobs.

PDB checker:
Fixed casting bug when the occupancy field is missing entirely from the PDB - example case was the live server job 1887/1888.

Javascript:
Inform the user if 'Check form' is/is not successful.
Replaced Javascript todos with tags referenced in the trac system.
If len pdbid != 4 then make sure it starts with an @.
Fixed highlighting bug in SK demo values when we corrupt the residues for testing - due to missing numSeqTolSK = 5.



git-svn-id: file:///kortemmelab/shared/commonSVN/rosettaweb@335 fc5780ad-6544-46e6-b60b-86c75d3988a5

Former-commit-id: 5f455205127f76dc7792cc104facae0f850706b5