Added working ddgdaemon code (and split it from rosetta_daemon).
Added working code to RosettaTasks.py to be removed in the next commit.

Hopefully fixed the bug when numtasks == 1 in ClusterScript/ClusterTask. Added ddgClusterScript.

ddG:
Added reconnection code for the databases in case of error 2006 (MySQL server went away) or 2013 (seemingly more serious but fixed with a reconnect). I need to investigate why these are necessary.

common:
Added rosettahelper.readFileLines() to return a list of lines from a file.


git-svn-id: file:///kortemmelab/shared/commonSVN/rosettaweb@579 fc5780ad-6544-46e6-b60b-86c75d3988a5

Former-commit-id: 3cb77ba06c8ef90a335b3cbd086ac64851cc8da2