Fixes to get PDB files with only CA atoms to work properly with the relatrix.
I added the chain type 'Protein skeleton' to the PDB class which identifies when only CA atoms are present in the PDB chain.
The PDBMissingMainchainAtomsException exception in pdb.py is raised by PDB.construct_pdb_to_rosetta_residue_map if skeletal chains exist in the PDB as Rosetta will throw an exception in this case.
I added a test case for 1LRP which has this problem.


Former-commit-id: 01b4daa56883ddd68ff285c35d1eb811c67567e2