Adding a preliminary stab at a sqlt wrapper. The class needs to be parameterized (hardcoded values at present).
There's a weird bug where sqlt-diagram fails if you wait for the subprocess or sleep inside _generate_schema_diagram but it does not fail if you sleep in generate_schema_diagram. I'm guessing that the subprocess really does fail but because we don't wait, the object gets garbage-collected when the function quits and the script continues on. Despite the fact that the sqlt-diagram call fails, an image is still created using the current approach.


Former-commit-id: daa4aebac628310d1fc0909ee94f73fd404eadb0