Generalized some of the alignment classes.
The existing SequenceAlignmentPrinter only handles two sequences. The new MultipleSequenceAlignmentPrinter handles multiple sequences.

Similarly, PDBChainMapper only provides mapping functionality for a pair of related PDBs e.g. Scaffold and Model. The new PipelinePDBChainMapper provides mapping functionality for an arbitrary sequence of PDBs e.g. Scaffold, Model, and Design although these PDBs are assumed to be closely related. A useful function (and the impetus for this) is get_differing_residue_ids where you can ask for all residues which differ from one PDB e.g. 'Design' and any other PDBs e.g. 'Scaffold' and 'Model'. This lets you get a list of residues that you can highlight in a e.g. PyMOL session. As with PDBChainMapper, PipelinePDBChainMapper can create sequence alignments strings in plaintext or HTML for web services.

ScaffoldModelDesignChainMapper now inherits from PipelinePDBChainMapper. This fixed a deficiency in ScaffoldModelDesignChainMapper where there was no functionality to get the list of residues of the design which differed from the model or scaffold (in the PyMOL sessions, I was incorrectly reusing the model's list for this purpose which only works in the special cases where the model residue IDs match the design residue IDs).

In the clustalo module, I enabled the code for multiple sequences. It worked fine - I had only disabled it as I did not have a test case when I ran that code initially.

Bugfix: Removed a weird tautology in SequenceAlignmentPrinter (len(self.seq1_name) == len(self.seq1_name)). seq2_name was probably intended here but the expression does not seem to make sense in that case either.

Todo: PipelinePDBChainMapper has not been tested in cases where chain A in pdb1 maps to chains A and B in pdb2. I don't think the code will stand up here. A good test case would be 3MWO -> 1BN1 where 3MWO:A and 3MWO:B both map to 1BN1:A

Todo: Update the code in the ScaffoldModelCrystalBuilderPyMOL session class. I may want to add functionality to this class to pass many lists of residues rather than the three sets of differing residues I currently pass. This would help for ticket 1149 (residues of interest) where we want to highlight motif residues, residues important for stability etc.

Former-commit-id: 0b36fefe92a03241bd79bed5eff26deed19187b5