The ScaffoldModelDesignBuilder now creates objects as well as selections so the user can easily turn on and off the e.g. mutated residues.
Renamed 'crystal' to 'design' in the PyMOL classes and alignment.py.
Added a settings argument to the PyMOL sessions. At present, this is only used to set a background color in ScaffoldModelDesignBuilder but it could potentially be more useful. One use of this would be to allow users to customize the PyMOL sessions based on their color preferences.

API change: generate_pymol_session now returns a PyMOL session and the script which created it as opposed to just the session.