Adding a test case for the X edge-case for the scaffold, model, design PSE generation.
The problem here is that the scaffold and design both have two chains whereas the model has one chain however all the chains are pretty much the same chain. The chains are related in an X pattern (2-to-1 then 1-to-2 as we go from scaffold to model to design). One solution would be to add the reverse mappings i.e.
   - the scaffold-model mapping is (A->A, B->A), the model-scaffold mapping is (A->A)
   - for each pair in the scaffold-model mapping, add the opposite mapping to the model-scaffold mapping if it does not already exist. This results in a model-scaffold mapping of (A->A, A->B).
Since the mappings are based on the percentage identity cutoff and should be symmetric relations (I believe), this seems reasonable enough. It would be good to dig into the code to see why I do not generate a relation (A->A, A->B) rather than (A->A) for the model-scaffold mapping so I'll do that first.