I split get_sequence_alignment_strings into a subfunction get_sequence_alignment_printer_objects which returns MultipleSequenceAlignmentPrinter objects. get_sequence_alignment_strings_as_html now calls subfunction get_sequence_alignment_printer_objects instead which avoids a bunch of unnecessary string parsing. Moved the HTML generation out of the PipelinePDBChainMapper class and into the MultipleSequenceAlignmentPrinter class where it makes sense to live. ScaffoldModelChainMapper updated to use PipelinePDBChainMapper. PDBChainMapper has been removed.
API change: generate_pymol_session now returns a pair: a PSE file and the script that created it.

Bugfix: Fixed a bug in HTML generation that would mark up missing parts of sequence e.g. if the Scaffold and Model had an alanine at position x but the Design had missing information there and had a '-', we would incorrectly mark up the Scaffold and Model residues. This is related to a bug in the PSE generation for the orthoCdc42:orthoITSN interface case.
Bugfix: In the PipelinePDBChainMapper constructor ('self.chain_mapping[mapping_key][pdb2_chain_id] = mcl').

Former-commit-id: e57e26c385c64f93ebced1240b019be0cd571316