Adding new PDB object functions.
get_annotated_chain_sequence_string returns a Sequence object for either the SEQRES or ATOM sequences. If a SEQRES sequence is requested, it is returned when available, otherwise an ATOM sequence is returned. The return value is a tuple which states the type of sequence ('SEQRES' or 'ATOM'). get_chain_sequence_string is a related function.

construct_seqres_to_atom_residue_map does exactly what it says on the tin. The mapping is done via the SequenceAligner i.e. via Clustal at present. The ResidueRelatrix should be used in preference to this function but that it not always possible (PDB has not been deposited or has no SIFTS record) so this function serves a purpose.


Former-commit-id: 388b4be5725297c28616533d3179c3f183ea0671