An important high-level change to the PipelinePDBChainMapper class (and so the ScaffoldModelChainMapper class).
For reasons motivated by the Designs website, we now perform sequence alignments based on the SEQRES sequences if they exist (if not we fall back to the ATOM sequences). We need the differing ATOM residue IDs for the PyMOL session so we take the sequence alignments and convert the differing residue IDs into ATOM residues via the new PDB.construct_seqres_to_atom_residue_map function. This new behavior is optional via the use_seqres_sequences_if_possible parameter but it is the default behavior.

To be more explicit, the motivating case was aligning 1KI1 to 3QBV where ClustalO gave a bad alignment by default. We switched to ClustalW in a previous commit but it still makes more sense to align the SEQRES sequences. The mapping from SEQRES to ATOM sequences then gives us the information we need for the PyMOL sessions.

match_pdb_chains, which returns a mapping from the chain IDs in one PDB to the related-by-sequence chain IDs in the other, now takes an argument use_seqres_sequences_if_possible which will use the SEQRES sequences to construct this chain to chain mapping.

Bugfix: ClustalW does not, by default, respect the order of the input sequences. We rely on this order being maintained in our parsing logic. I added the -OUTORDER=INPUT argument to ensure this (rather than change the parsing logic :-)).

Former-commit-id: 63eb79f51be5b82c8598da57710a2712eed2cc1c