Added tooltips for the HTML generation. If a residue in a sequence (a sequence may be a SEQRES sequence or an ATOM sequence) has an ATOM record, we show a tooltip with the corresponding ATOM residue id. This can be used for lots of fancy functions including my idea on how to implement the "residues of interest" functionality.
Issues: The tooltips are not 100% at present. For example, in the 1KI1 -> model -> 3QBV case, in chain B of the alignment from Scaffold to Model and ExpStructure, the tooltips in the Scaffold and ExpStructure start to become erroneous. This seems to happen after TYR D1513, TYR B1513, TYR B1513 in the scaffold, model, and exp. structure respectively. The exp. structure has an inserted residue which seems to mess up the mapping. I will fix this in a future commit as well as show the residue type in the tooltip (like the PDB website) so that this problem is more obvious.

Bugfix: Fixed get_sequence_alignment_strings_as_html (using extend instead of append since that function should return a string).