Ticket 1174.
Adding a non-strict argument to PDB.
This argument lets PDB files with unidentified residues pass. The motivation for this was with a bad PDB file for design BH32 in the designs database but it is a typical requirement.

Improvements to the aligner:
Added a non-strict argument to the aligner which is propagated to PDB objects created therein.
The PDBChainMapper classes and the match_RCSB_pdb_chains function now catch particular PDB errors and raise a PDBChainMapperException.
We now automatically remove sequences which contain all 'X' residues since Clustal will fail with those anyway.
Fixed a bug in match_pdb_chains - removal from a list while iterating along the list (facepalm).

e.g.
>>> a=[1,2,3,4]
>>> for c in a:
...   a.remove(c)
...
>>> print(a)
[2, 4]