Remove the pdb module's biopython requirement.
Biopython was being used to get at the BLOSUM and PAM matrices.  I
wrapped the import statement in a try/catch block so it fails gracefully
when biopython is missing.  This is important to me because I want to
use the pdb module to convert structures to fasta files on the cluster,
where biopython is not installed (and cannot easily be installed).

I also tweaked to pdb-to-fasta code so that it falls back on the atom
sequence if there is no seqres sequence.