NOTE: Major change to the returned data format. We now use a more hierarchical format which respects the fact that a PDB chain can have multiple SCOP and Pfam domains. BROKEN: The changes to the format break the CSV and table functions. I need to fix these but the paper is the priority at the moment. ALMOST WORKING: This code was initially meant to return the per-chain SCOP and Pfam information but I switched focus to the per-residue SCOP and Pfam information since that is what we needed for the paper. There is a little work to go finishing up the per-chain information (including fixing the CSV and table formats). See the print statements that I left in.
Adding determine_SCOPe_class_of_pdb_residue and determine_Pfam_class_of_pdb_residue functions. NOTE: I discovered later that it is easier to get this information from the residue records in the SIFTS files but this will be replaced in the future with the SIFTS database lookup which should be quicker.
Adding residue range field.
Ignoring root level in SCOP hierarchy.