Splitting old notion of "ligand" into three types: ligands e.g. GFP, ions e.g. Zn 2+, and solution e.g. HOH, DOD. This makes a lot more sense both in terms of storing the data in a database and separating out the information in which users will be interested. This will also be useful to handle mappings between RCSB PDB files and PDB files where the chain identities and ligand sequence IDs have been changed for arbitrary reasons (e.g. changing FPP on chain B to LIG on chain X in some computational pipeline) as we will typically only care about mappings from ligand IDs to RCSB ligand IDs; it seems less likely that ion or solution sequence IDs would be modified.
We now inspect HETATM records to find ligands, ions, and solution. This is important not only for solution -  which is typically omitted from the HET section - but also for PDBs output from some tool or suite e.g. Rosetta or which have been manually edited by users and are missing headers. The information in the HETATM records is limited to the determined coordinates but typically we care about the ligands/ions present rather than all atoms of those ligands or ions.

Adding a new chain type, "Heterogen" for when e.g. a "chain" contains only solution and/or ions.