Adding a very rough analysis filter function.
todo:
Filtering should be moved into the analysis stage and be based on the dataframe.
Also, we should not need two branches with essentially the same code as I have written here - the restriction and removal criteria should be combined in the filtering function to create one set of properties so that we use 'Buried' or 'Exposed' but not both. Since analysis will probably focus on a small number of properties, it is probably more practical to build the removal set i.e. the properties which are used to filter out records i.e. if 'Buried' is in the set of properties, we ignore any buried records.