Adding class to map ligand IDs between PDB files. This is useful for when a user has modified the chain, residue ID, and/or the ligand code for a ligand and you want to relate this back to the original RCSB file.
Use case:
The user renames 'GDP' to 'LIG' and changes the chain from 'C' to 'X' and the residue ID from '1250A' to '   1 '. We want to look up properties of the ligand in some automated pipeline.

Solution:
Using this mapping, you retrieve the original RCSB code. Ligand information can then be retrieved using other functions in this class. The original residue ID can be used to inspect the RCSB file to look for potential artifacts or b-factors etc.