Starting work on a new sequence aligner which takes advantage of structures with identical ATOM sequences to reduce the amount of matching. We use an equivalence relation over the ATOM sequences and only map representatives of each equivalence class.
Use case: creating a mapping from 50 wildtype structures with the same sequence to 50 mutant structures (usually where the mutant sequences are identical to each other).
In the old code, we would do O(n^2) alignments where n is the total number of structures. In the new code, we will do O(c^2) alignments where c is the number of equivalence classes and c <= n.