Adding another CADRES plot - the correctness "abacus" plot.
The correctness "abacus" plot displays measurements M (e.g. the difference in binding affinity given in kcal/mol)
over a set of cases (e.g. peptide binders) and possibly many categories (e.g. binding partners). For each case on the X-axis,
the measurement is shown on the Y-axis as a point.

Points are further distinguished by a second set of measurements N where |M| = |N|. Each m in M is categorized into one of
three classes: beneficial, neutral, or deleterious according to some cut-off for M. If the corresponding n in N has the same
categorization according to the cut-off for N then this point is plotted as a grey circle indicating a correct prediction.
Otherwise, the point is plotted as:
  - a red triangle if n is classified as deleterious;
  - a blue diamond if n is classified as neutral;
  - a green triangle if n is classified as beneficial.

This plot allows incorrectly classified values n in N to be distinguished by color and shape and also indicates the type
of wrong classification. A shaded region is drawn on the plot to identify the neutral zone which helps to identify cases
which are near-misses due to edge artifacts.

This function was used to generate assessment plots for CADRES 2016.
See .examples/plot/graphs for expected output.